packages/pykit-dataset/pyproject.toml
[project]
name = "pykit-dataset"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pykit_dataset"]

[project.optional-dependencies]
huggingface = [
    "datasets",
    "huggingface-hub",
    "Pillow",
]
kaggle = [
    "kagglehub",
]
web = [
    "httpx",
    "Pillow",
]
all = [
    "datasets",
    "huggingface-hub",
    "kagglehub",
    "httpx",
    "Pillow",
]
